removed unused APIs and renamed existing
