+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/compose.py
@@ -12,13 +12,13 @@ from ..shared_store import SharedStore
 from ..utils import InvalidComposeSpec
 
 logger = logging.getLogger(__name__)
-compose_router = APIRouter()
+compose_router = APIRouter(tags=["docker-compose"])
 
 
 @compose_router.post(
     "/compose:store", response_class=PlainTextResponse, responses={204: {"model": None}}
 )
-async def store_docker_compose_spec_for_later_usage(
+async def validates_docker_compose_spec_and_stores_it(
     request: Request, response: Response
 ) -> Optional[str]:
     """ Expects the docker-compose spec as raw-body utf-8 encoded text """
@@ -37,43 +37,8 @@ async def store_docker_compose_spec_for_later_usage(
     return None
 
 
-@compose_router.post("/compose:preload", response_class=PlainTextResponse)
-async def create_docker_compose_configuration_containers_without_starting(
-    request: Request, response: Response, command_timeout: float
-) -> str:
-    """ Expects the docker-compose spec as raw-body utf-8 encoded text """
-    body_as_text = (await request.body()).decode("utf-8")
-
-    settings: DynamicSidecarSettings = request.app.state.settings
-    shared_store: SharedStore = request.app.state.shared_store
-
-    try:
-        shared_store.put_spec(body_as_text)
-    except InvalidComposeSpec as e:
-        logger.warning("Error detected %s", traceback.format_exc())
-        response.status_code = HTTP_400_BAD_REQUEST
-        return str(e)
-
-    # --no-build might be a security risk building is disabled
-    command = (
-        "docker-compose --project-name {project} --file {file_path} "
-        "up --no-build --no-start"
-    )
-    finished_without_errors, stdout = await write_file_and_run_command(
-        settings=settings,
-        file_content=shared_store.compose_spec,
-        command=command,
-        command_timeout=command_timeout,
-    )
-
-    response.status_code = (
-        HTTP_200_OK if finished_without_errors else HTTP_400_BAD_REQUEST
-    )
-    return stdout
-
-
 @compose_router.post("/compose", response_class=PlainTextResponse)
-async def start_or_update_docker_compose_configuration(
+async def runs_docker_compose_up(
     request: Request, response: Response, command_timeout: float
 ) -> str:
     """ Expects the docker-compose spec as raw-body utf-8 encoded text """
@@ -99,7 +64,7 @@ async def start_or_update_docker_compose_configuration(
 
 
 @compose_router.get("/compose:pull", response_class=PlainTextResponse)
-async def pull_docker_required_docker_images(
+async def runs_docker_compose_pull(
     request: Request, response: Response, command_timeout: float
 ) -> str:
     """ Expects the docker-compose spec as raw-body utf-8 encoded text """
@@ -136,39 +101,8 @@ async def pull_docker_required_docker_images(
     return stdout
 
 
-@compose_router.put("/compose:stop", response_class=PlainTextResponse)
-async def stop_containers_without_removing_them(
-    request: Request, response: Response, command_timeout: float
-) -> str:
-    """Stops the previously started service
-    and returns the docker-compose output"""
-    shared_store: SharedStore = request.app.state.shared_store
-    settings: DynamicSidecarSettings = request.app.state.settings
-
-    stored_compose_content = shared_store.compose_spec
-    if stored_compose_content is None:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return "No started spec to stop was found"
-
-    command = (
-        "docker-compose --project-name {project} --file {file_path} "
-        "stop --timeout {stop_and_remove_timeout}"
-    )
-    finished_without_errors, stdout = await write_file_and_run_command(
-        settings=settings,
-        file_content=stored_compose_content,
-        command=command,
-        command_timeout=command_timeout,
-    )
-
-    response.status_code = (
-        HTTP_200_OK if finished_without_errors else HTTP_400_BAD_REQUEST
-    )
-    return stdout
-
-
 @compose_router.delete("/compose", response_class=PlainTextResponse)
-async def remove_docker_compose_configuration(
+async def runs_docker_compose_down(
     request: Request, response: Response, command_timeout: float
 ) -> str:
     """Removes the previously started service
