--- services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/container.py
@@ -1,100 +1,0 @@
-from typing import Any, Dict, Union
-
-import aiodocker
-from fastapi import APIRouter, Query, Request, Response
-from starlette.status import HTTP_400_BAD_REQUEST
-
-from ..shared_store import SharedStore
-
-container_router = APIRouter()
-
-
-@container_router.get("/container/{name_or_id}/logs")
-async def get_container_logs(
-    # pylint: disable=unused-argument
-    request: Request,
-    response: Response,
-    name_or_id: str,
-    since: int = Query(
-        0,
-        title="Timstamp",
-        description="Only return logs since this time, as a UNIX timestamp",
-    ),
-    until: int = Query(
-        0,
-        title="Timstamp",
-        description="Only return logs before this time, as a UNIX timestamp",
-    ),
-    timestamps: bool = Query(
-        False,
-        title="Display timestamps",
-        description="Enabling this parameter will include timestamps in logs",
-    ),
-) -> Union[str, Dict[str, Any]]:
-    """ Returns the logs of a given container if found """
-    shared_store: SharedStore = request.app.state.shared_store
-
-    if name_or_id not in shared_store.container_names:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=f"No container '{name_or_id}' was started")
-
-    docker = aiodocker.Docker()
-
-    try:
-        container_instance = await docker.containers.get(name_or_id)
-
-        args = dict(stdout=True, stderr=True)
-        if timestamps:
-            args["timestamps"] = True
-
-        container_logs: str = await container_instance.log(**args)
-        return container_logs
-    except aiodocker.exceptions.DockerError as e:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=e.message)
-
-
-@container_router.get("/container/{name_or_id}/inspect")
-async def container_inspect(
-    request: Request, response: Response, name_or_id: str
-) -> Dict[str, Any]:
-    """ Returns information about the container, like docker inspect command """
-    shared_store: SharedStore = request.app.state.shared_store
-
-    if name_or_id not in shared_store.container_names:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=f"No container '{name_or_id}' was started")
-
-    docker = aiodocker.Docker()
-
-    try:
-        container_instance = await docker.containers.get(name_or_id)
-        inspect_result: Dict[str, Any] = await container_instance.show()
-        return inspect_result
-    except aiodocker.exceptions.DockerError as e:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=e.message)
-
-
-@container_router.delete("/container/{name_or_id}/remove")
-async def container_remove(
-    request: Request, response: Response, name_or_id: str
-) -> Union[bool, Dict[str, Any]]:
-    shared_store: SharedStore = request.app.state.shared_store
-
-    if name_or_id not in shared_store.container_names:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=f"No container '{name_or_id}' was started")
-
-    docker = aiodocker.Docker()
-
-    try:
-        container_instance = await docker.containers.get(name_or_id)
-        await container_instance.delete()
-        return True
-    except aiodocker.exceptions.DockerError as e:
-        response.status_code = HTTP_400_BAD_REQUEST
-        return dict(error=e.message)
-
-
-__all__ = ["container_router"]
