+++ services/dynamic-sidecar/src/simcore_service_dynamic_sidecar/api/mocked.py
@@ -10,35 +10,35 @@ from fastapi import APIRouter
 
 logger = logging.getLogger(__name__)
 
-mocked_router = APIRouter()
+mocked_router = APIRouter(tags=["Mocked frontend calls"])
 
 
 @mocked_router.post("/push")
-async def post_push() -> str:
+async def ignored_push_post() -> str:
     logger.warning("ignoring call POST /push from frontend")
     return ""
 
 
 @mocked_router.get("/retrieve")
-async def get_retrive() -> str:
+async def ignored_port_data_load() -> str:
     logger.warning("ignoring call GET /retrive from frontend")
     return ""
 
 
 @mocked_router.post("/retrieve")
-async def post_retrive() -> str:
+async def ignored_port_data_save() -> str:
     logger.warning("ignoring call POST /retrive from frontend")
     return ""
 
 
 @mocked_router.get("/state")
-async def get_state() -> str:
+async def ignored_load_service_state_state() -> str:
     logger.warning("ignoring call GET /state from frontend")
     return ""
 
 
 @mocked_router.post("/state")
-async def post_state() -> str:
+async def ignored_save_service_state_state() -> str:
     logger.warning("ignoring call POST /state from frontend")
     return ""
