+++ services/dynamic-sidecar/tests/unit/test_api_compose.py
@@ -49,25 +49,7 @@ async def test_store_compose_spec_invalid(test_client: TestClient) -> None:
     assert len(response.text) > 28
 
 
-async def test_preload(test_client: TestClient, compose_spec: Dict[str, Any]) -> None:
-    response = await test_client.post(
-        f"/{api_vtag}/compose:preload",
-        query_string=dict(command_timeout=5.0),
-        data=compose_spec,
-    )
-    assert response.status_code == 200, response.text
-
-
-async def test_preload_compose_spec_not_provided(test_client: TestClient) -> None:
-
-    response = await test_client.post(
-        f"/{api_vtag}/compose:preload", query_string=dict(command_timeout=5.0)
-    )
-    assert response.status_code == 400, response.text
-    assert response.text == "\nProvided yaml is not valid!"
-
-
-async def test_compuse_up(
+async def test_compose_up(
     test_client: TestClient, compose_spec: Dict[str, Any]
 ) -> None:
     # store spec first
@@ -108,18 +90,7 @@ async def test_pull_missing_spec(
     assert response.text == "No started spec to pull was found"
 
 
-async def test_stop_missing_spec(
-    test_client: TestClient, compose_spec: Dict[str, Any]
-) -> None:
-    response = await test_client.put(
-        f"/{api_vtag}/compose:stop",
-        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
-    )
-    assert response.status_code == 400, response.text
-    assert response.text == "No started spec to stop was found"
-
-
-async def test_compuse_stop_after_running(
+async def test_compose_delete_after_stopping(
     test_client: TestClient, compose_spec: Dict[str, Any]
 ) -> None:
     # store spec first
@@ -130,34 +101,6 @@ async def test_compuse_stop_after_running(
     # pull images for spec
     response = await test_client.post(
         f"/{api_vtag}/compose",
-        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
-    )
-    assert response.status_code == 200, response.text
-
-    response = await test_client.put(
-        f"/{api_vtag}/compose:stop",
-        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
-    )
-    assert response.status_code == 200, response.text
-
-
-async def test_compuse_delete_after_stopping(
-    test_client: TestClient, compose_spec: Dict[str, Any]
-) -> None:
-    # store spec first
-    response = await test_client.post(f"/{api_vtag}/compose:store", data=compose_spec)
-    assert response.status_code == 204, response.text
-    assert response.text == ""
-
-    # pull images for spec
-    response = await test_client.post(
-        f"/{api_vtag}/compose",
-        query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
-    )
-    assert response.status_code == 200, response.text
-
-    response = await test_client.put(
-        f"/{api_vtag}/compose:stop",
         query_string=dict(command_timeout=DEFAULT_COMMAND_TIMEOUT),
     )
     assert response.status_code == 200, response.text
